+++ src/etcd/src/ConfigProvider.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
+
+namespace Hyperf\Etcd;
+
+class ConfigProvider
+{
+    public function __invoke(): array
+    {
+        return [
+            'dependencies' => [
+                KVInterface::class => KVFactory::class,
+            ],
+            'commands' => [
+            ],
+            'scan' => [
+                'paths' => [
+                    __DIR__,
+                ],
+            ],
+            'publish' => [
+                [
+                    'id' => 'config',
+                    'description' => 'The config for etcd.',
+                    'source' => __DIR__ . '/../publish/etcd.php',
+                    'destination' => BASE_PATH . '/config/autoload/etcd.php',
+                ],
+            ],
+        ];
+    }
+}